split new legend into multiple files
